refactor(config, repo): all writes go through FSRepo. next: privatize these
