+++ repo/config/version_test.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"encoding/json"
 	"strings"
 	"testing"
 )
@@ -23,8 +24,7 @@ func TestAutoUpdateValues(t *testing.T) {
 	}
 
 	for i, tc := range tests {
-		err := Decode(strings.NewReader(tc.input), &tval)
-		if err != tc.err {
+		if err := json.NewDecoder(strings.NewReader(tc.input)).Decode(&tval); err != tc.err {
 			t.Fatalf("%d failed - got err %q wanted %v", i, err, tc.err)
 		}
 
@@ -32,5 +32,4 @@ func TestAutoUpdateValues(t *testing.T) {
 			t.Fatalf("%d failed - got val %q where we wanted %q", i, tval.AutoUpdate, tc.val)
 		}
 	}
-
 }
