+++ repo/fsrepo/serialize_test.go
@@ -1,13 +1,15 @@
-package config
+package fsrepo
 
 import (
 	"testing"
+
+	config "github.com/jbenet/go-ipfs/repo/config"
 )
 
 func TestConfig(t *testing.T) {
 	const filename = ".ipfsconfig"
 	const dsPath = "/path/to/datastore"
-	cfgWritten := new(Config)
+	cfgWritten := new(config.Config)
 	cfgWritten.Datastore.Path = dsPath
 	err := WriteConfigFile(filename, cfgWritten)
 	if err != nil {
